requirements.txt
click>=8.1.0
inquirer>=3.1.0
GitPython>=3.1.0
PyGithub>=2.0.0
colorama>=0.4.6
pydantic>=2.0.0